Lifting the detection of invalid import placement up to the parser
